📦 NEW: Profile button in navbar
