+++ client/src/components/Navigation/MobileNavBar/MobileNavBar.css
@@ -25,3 +25,12 @@
 .logo img {
   margin-right: 5px;
 }
+
+.profilepic {
+  height: 36px;
+  color: #f0f2f5;
+  position: absolute;
+  right: 1rem;
+  border-radius: 50%;
+  cursor: pointer;
+}
